[consensus] move RocksDBStore to use DBMapUtils
## Description 

* switches consensus::RocksDBStore to DBMapUtils (as used in the rest of
the codebase) as a preparation step for experiments with TideHunter
* deprecates the `typed_store::reopen!` macro and its remaining call
sites.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: