[replay] Update effects checking to handle the new abort_error field
## Description 

The saved `HaneulTransactionBlockEffects` used for the transaction replay
CI job do not contain this new field. This leads to replay failures.

This updates the replay checking logic to ignore the new field if the
old saved effects `abort_error` field is `None` (i.e., was not present
in the serialized sandbox snapshot).

## Test plan 

Ran the sandbox CI tests locally.