[haneul][cli] Support generating package summaries from the Haneul CLI
## Description

This adds support for generating package summaries from the Haneul CLI.
There are two main ways summaries can be generated:

1. Running `haneul move summary` in a source Move package. This will
generate summaries for the current Move package along with its
dependencies. The generated summaries will contain source annotations
(e.g., doc comments, annotations, parameter names etc).
2. Running `haneul move summary --package-id <object-id> --bytecode`
anywhere. This will generate a bytecode-based summary of the package at
`<object-id>` along with all of its dependencies. Note that this will
not contain source annotations (doc comments, annotations, parameter
names etc.) as none exist in this case.

You can specify a specific output directory to be used for the generated
summaries with the `--output-directory <path>` flag. If none is provided
the output directory of `package_summaries` is chosen, and this will be
located either at the root of the Move package, or under the directory
where the command was run.

You may currently specify one of two supported output formats today --
either `json` or `yaml` with `-o yaml` or `-o json`. The default chosen
is `json`. All files, in the generated output will respect this format
flag (both per module and the package metadata below).

Additionally, an `address_mapping.{json, yaml}` is generated in the root
of the generated summaries. In the case of a source package this is a
mapping of the literal address value to the name of the address. In the
case of bytecode-based summaries, this is an identity mapping (and does
_not_ contain linkage information!).

For Haneul CLI usage specifically, an additional,
`root_package_metadata.{json, yaml}` file is generated that contains
metadata information based on the package supplied -- e.g., linkage
table (and version for every package), type origins for every package
etc. This metadata is empty for any source-generated summary as we don't
know what network to generate this information for. In the future with
environments we could look into generating this information from the
lockfile, but for now the metadata is not filled in in this case.

## Test plan 

Added new shell tests for this