display(v2): parsing byte literals
## Description

Byte literals store the same data as a string literal, but exposes it as
the raw bytes (i.e. it has type `vector<u8>`) rather than as a string.

This also requires a use of the 2 token lookahead, to make sure that the
token after the `b` keyword is a string literal (otherwise it should be
an identifier).

## Test plan

```
haneul$ cargo nextest run -p haneul-display
```