Protocol update to allow recovery of stolen funds
This Pull Request implements a path to recover stolen Cetus funds from
the two hacker addresses, in the event that the community vote passes.

The recovery is implemented as follows:

- ProtocolConfig can specify address aliases which allow a signature
from address A to act as the sender address B.
- The aliasing is restricted to a specified set of TransactionDigests.
Only those exact transactions may be aliased.
- The address deny list is bypassed for any allowed transaction.

The exact transaction digests and addresses are TBD, but in the event
the governance vote passes, the transactions will be made available for
inspection by validators and the community prior to doing the protocol
upgrade.