display(v2): More permissive whitespace detection
## Description

Previously the lexer outputted an explicit whitespace token, this was so
that string and numeric literal patterns could detect two tokens next to
each other, but it made other parts of the grammar less permissive. In
particular, it affected parsing for datatype literals, which start with
a struct tag.

To distinguish a numeric literal from the start of a data type, the
grammar needed to perform a lookahead operation. For simplicity and
because of the presence of explicit whitespace tokens, the lookahead was
implemented to look for an immediately following '::' token (so did not
allow for whitespace) and the rest of the struct tag and datatype parser
did not permit whitespace, for consistency reasons.

The revised implementation is more permissive -- it allows us to detect
tokens that are immediately next to each other, and also permit
whitespace between tokens during lookahead.

It works by introducing a boolean to each lexeme which indicates whether
the lexeme was preceded by whitespace (fusing a previous whitespace
token with the real token).

This means we can implement lookahead without sacrificing support for
whitespace between struct tag and enum literal. It also reduces the
opportunity for errors due to omitting calls to `eat_whitespace()`.

## Test plan

Updated unit tests:

```
haneul$ cargo nextest run -p haneul-display
```