[cli] Add `sender` to `ptb` help menu
## Description 

A previous PR  added support for setting the `sender` in a
transaction. This PR adds this new argument to the ptb help menu.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Added a new `--sender` argument for transactions to set the
sender to a specific address.
- [ ] Rust SDK: