[3/n][object runtime type tags] Add type tags to object runtime and update adapter to handle them.
## Description 

Re-land of the type-tags in object runtime PR -- with some changes to
utilize the new functionality in . The new code changes are all in
the top commit (`d7b8fbd6845b3398f52f726c7952ca8ec3d8f747`).

This utilizes the new `try_load_cached_type` function in the VM runtime
that was added in  to load a type from a typetag, that we know
must be in VM's cache since it is part of the results of the transaction
(either as an event, as a mutated or create object, or because we are in
test scenario and we have a single cache for the whole test).

## Test plan 

Existing tests + a couple new ones as well as tested the CI/sandbox
tests locally and verified they all passed.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [X] Protocol: Changes to using type tags in the object runtime. No
user-visible impacts.
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------