Plumbing for accumulator events
This PR contains most of the plumbing from , which can be merged
more or less as-is.

Since this doesn't do anything on its own yet, the motivation may not be
clear in all cases, but  can provide the context.

(Also, there are no tests yet because, again, this doesn't do anything
yet.)