[static-ptbs] Add start of invariant checking passes
## Description 

Adds invariant passes that run post typing in the new static PTBs.

This adds a simple defining-id checker for types, but more advanced ones
can/will be coming, but this is mainly to get the plumbing in to do it.

There was a question in my mind of if we wanted to build up a more
advanced analysis framework over the typed PTBs, but I decided against
that is most likely premature over-engineering (esp. given the
simplicity in the PTB AST).

## Test plan 

😍