[rpc-alt] report name of unknown methods in metrics too
## Description 

This way we know definitively from metrics which methods we need to add.
This is only temporary though as we push for migration of first party
apps. We will have to revert if we ever open up to the public.

## Test plan 

existing test

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: