[consensus] Add new shared consensus types crate
## Description 

This crate will be used for shared types between haneul and consensus that
can be used as a lighter dependency and also stop us from recreating the
type on the haneul side. This is just the first PR, will move more shared
types into this crate in future PRs.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: