Only spawn metric task if reporting is enabled
## Description

This morning Michael and I shut off the pipes for the legacy stack,
which caused me to get a few lag alerts. Currently we can "disable" this
metric, but we still emit a -1 value when it is disabled. I am changing
this behavior to just completely not emit the metric instead, which will
silence the alerts.