[cli] Fix `haneul move build --dump-bytecode-as-base64` to correctly take into account Move.lock addresses
## Description 

This PR fixes a bug where `haneul move build --dump-bytecode-as-base64`
uses a published address in the resolved table for the root package
being built instead of `0x0`, leading to inconsistencies when this
interacts with automated feature management.

## Test plan 

👀 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: `haneul move build --dump-bytecode-as-base64` is now working
correctly due to a bug in the logic when a Move.lock file existed with
the published address.
- [ ] Rust SDK: