graphql: fix graphiql dependency
## Description

GraphiQL is currently not working because of an issue with `unpkg`
knowing that `graphiql` `5.0.0` has been released but not having the
resources for it.

The fix is to cherry-pick some upstream changes to `async-graphql` that
pins the version of GraphiQL to use.

## Test plan

Run the GraphQL service locally -- previously the UI would not load, and
after this change it does.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: