[replay] Add option to build packages preserving their on-chain ID
## Description 

When compiling packages locally (but NOT for publish/upgrade) their
module IDs are populated with `0x0`. This is required for these modules
to be publishable/upgradeable but may be problematic in some
circumstances. For example, when generating traces from on-chain
transactions, function call targets refer to an actual on-chain ID of a
package containing these functions. Trace analyzer wishing to match
locally build source code will fail to do so if the source code build
artifacts (such as debug info) contain `0x0` as package IDs

This PR adds the ability to build a package locally preserving its
on-chain ID in the locally built modules