[haneul-replay-2] Produce artifacts as result of replay
## Description 

Does a bunch of cleanup and moving some things around in preparation for
incorporating it into the the main Haneul CLI.

One of the main things is the addition of `Artifacts` these are on-disk
data that are the result of replay. The `ArtifactManager` is a way to
easily deal with these artifacts (finding them, reading them, writing to
them etc.) and more artifact types I'm sure will be added over time.

Existing artifacts (i.e., for the debugger) locations are not changed. 

## Test plan 

Ran some replays locally.

## TODO

Need to incorporate the saving the bytecode files and source maps out
via the artifact manager.