Add package_bcs table
## Description

Security is requesting a move_package_bcs table similar to
transaction_bcs. Currently they have to join to graphql or json rpc to
get the bcs which is complicated and error prone compared to just doing
it in snowflake.

I am deploying it as a separate instance of the indexer to do the
backfill and then I will merge this pipeline into the main instance once
the backfill is done.

## Test plan

It's running in k8s.