[Haneul CLI] Fixup feature flagging setting from Haneul CLI for replay with tracing
## Description 

We need to set the feature flag in the `haneul-replay-2` crate if the
`tracing` feature is set in the `haneul` crate.

## Test plan 

Ran locally 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: