[Haneul CLI] Build using specific networks lockfile if specified
## Description 

This uses the `--client.env` flag to allow building a package using a
specific network's address values.

Existing behavior of `haneul move build` is kept (and no network calls are
performed).

If built with a specified network -- e.g., `haneul move --client.env
testnet build` -- the address values for the `testnet` chain-id in the
lockfile are used. If an environment is specified a network request is
performed to get the specified environments chain ID.

## Test plan 

Tested locally