Move the sponsor to a set/unset api
## Description 

the `sponsor` field in `TxContextBuilder` is an option like `rgp` but
did not have a set/unset API.
Originally I did not care to do that too because the sponsor is truly an
`option` whereas the `rgp` is an `option` just to mean set it or ignore
it.
For consistency a `set/unset` API seems more proper, not sure what
others think, so this is the PR in case we want it. I'll cherry pick
tomorrow if accepted it

## Test plan 

changed the tests to use the set/unset API

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: