[mfp] Fix bug when re-sign executed transactions
## Description 

When a transaction has already been executed, but get voted on again, we
should always go through the signing flow.
Previously this is problematic because if a transaction has been
executed in a previous epoch, we would still vote for it in a new epoch,
causing a new certificate to be formed and coming out of consensus
again.
We could potentially fix it by checking whether it has been executed in
the current epoch.
But that may be premature optimzation.
keep it simple for now.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: