[pkg-alt] added implicit-deps field to manifest
## Description

This adds the `implicit-deps` field to the manifest format. There are 4
modes: enabled and disabled are clear. Legacy is only used for legacy
packages - for those, implicit deps are enabled unless any of them are
present. Finally, I added a "testing" mode as a way for us to do
implicit deps differently in the monorepo for our internal tests (this
is something that has been a headache with our existing implicit deps).

## Test plan 

I added unit tests. In principle these could be integration tests, but
my goal is to start migrating those back so I'm trying out the unit test
format.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------