use `debug_fatal` to track invariant violations
## Description 

`debug_fatal` raises an alert which is what we want for invariant
violation.
That is hooked to a slack channel for now so not very intrusive, which
is also good.
Not sure how to test this, which seems pretty safe but open to
suggestions

## Test plan 

YOLO? existing tests good enough?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] gRPC:
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: